requirements.txt
playwright>=1.48
pytest>=7.4
pytest-playwright
google-genai>=0.3
python-dotenv
requests